+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.5] - 2021-03-30
+- Removing dep on pinata gateway tools and cleaning up ipfs handling code
+- Changes for 1.2.4
+
+## [1.2.4] - 2021-03-30
+- [broken build] fix punks metadata fetching address
 
 ## [1.2.3]  - 2021-02-07
 - Fix timeout in example code